pyproject.toml
[project]
name = "cot-experiment"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]